requirements.txt
# ---------------------------
# Core Framework
# ---------------------------
fastapi
uvicorn
python-dotenv

# ---------------------------
# CrewAI + Tools
# ---------------------------
crewai==1.6.0
crewai-tools==1.6.0

# ---------------------------
# LLM Support (Groq + LiteLLM Router)
# ---------------------------
litellm
groq

# ---------------------------
# Document Generation 
python-docx
pdfplumber

streamlit


transformers


# ---------------------------
# Auth & Database
# ---------------------------
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
